backend/requirements.txt
numpy==1.19.5
opencv-python==4.1.2.30
sklearn
seaborn==0.11.1
scikit-image==0.18.1
tensorflow-gpu==2.4.1
keras==2.4.3
pandas==1.1.5
matplotlib==3.2.2